requirements.txt
pymongo
beautifulsoup4
bs4
requests
selenium
urllib3
pytz
python-dotenv
progress
PyJWT
bcrypt